+++ integration/upstream_client_test.go
@@ -269,23 +269,3 @@ func createSnapshotCache(updates int, logger log.Logger) (resourcev2.TestSnapsho
 		NumClusters:  updates,
 	}, gcpcachev2.NewSnapshotCache(false, gcpcachev2.IDHash{}, gcpLogger{logger: logger})
 }
-
-type gcpLogger struct {
-	logger log.Logger
-}
-
-func (logger gcpLogger) Debugf(format string, args ...interface{}) {
-	logger.logger.Debug(context.Background(), format, args)
-}
-
-func (logger gcpLogger) Infof(format string, args ...interface{}) {
-	logger.logger.Info(context.Background(), format, args)
-}
-
-func (logger gcpLogger) Warnf(format string, args ...interface{}) {
-	logger.logger.Warn(context.Background(), format, args)
-}
-
-func (logger gcpLogger) Errorf(format string, args ...interface{}) {
-	logger.logger.Error(context.Background(), format, args)
-}
